pyproject.toml
[project]
name = "boothlist"
version = "0.1.0"
description = "BoothList - BOOTH Asset Dashboard"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["beautifulsoup4>=4.14.3", "pyyaml>=6.0.3", "requests>=2.32.5"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["D", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
